docs/requirements.txt
furo
linkify-it-py
numpydoc
matplotlib
matplotlib-inline
mpmath
myst-parser
networkx
Sphinx
sphinx-autobuild
sphinx-copybutton
sphinx-math-dollar
sphinx-reredirects
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-jquery
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
sphinx-reredirects
chardet
Solverz
